# school-communication-platform/docs/openapi.yaml
openapi: 3.0.3
info:
  title: School Communication Platform API
  description: >
    REST API for five connected dashboards (Super Admin, Admin, Teacher,
    Student, Parent). Messaging, document sharing, announcements and
    notifications. Full reference: docs/API.md and /docs/api.html.
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /attendance/summary/student/{studentId}:
    get:
      summary: Attendance percentages + recent absences
      parameters:
        - { name: studentId, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: Success }
        '403': { description: Not authorized }
...
  /fees/student/{studentId}:
    get:
      summary: Fees + payments + balance (own/parent/admin)
      parameters:
        - { name: studentId, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: Success }
        '403': { description: Not authorized }
...